Create artist_stmnt_pos_parse
This files makes lists from parts of speech (pos) queries for use in possible poetry project. 
I would love some guidance on how to make the code more efficient in lines 70 - 94, I'm trying to select two word phrases that start with the word 'they' and are followed by a pos of 'VBP', and also remove the 'to be' verbs.